requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.20,<1.0
PyMuPDF>=1.26,<2.0
python-docx>=1.2,<2.0
PyYAML>=6.0,<7.0
numpy>=2.0,<3.0
openai>=1.0,<2.0
pytest>=8.0,<9.0
httpx>=0.28,<1.0
